requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.6.2.post1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
asyncpg==0.30.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
av==14.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloud-sql-python-connector==1.13.0
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
cryptography==43.0.3
ctranslate2==4.6.0
daphne==4.1.2
datasets==3.1.0
db-dtypes==1.4.2
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
Django==5.2
docopt==0.6.2
docstring_parser==0.16
durationpy==0.9
entrypoints==0.4
executing==2.1.0
fastapi==0.115.11
faster-whisper==1.1.1
fastjsonschema==2.21.1
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.1.0
flask-sock==0.7.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-speech==2.31.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jiwer==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
kubernetes==32.0.1
langchain==0.3.7
langchain-core==0.3.18
langchain-google-cloud-sql-pg==0.11.0
langchain-google-vertexai==2.0.7
langchain-text-splitters==0.3.2
langsmith==0.1.143
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.31.0
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.66.3
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
papermill==2.6.0
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
platformdirs==4.3.6
plotly==6.0.1
polling2==0.5.0
posthog==3.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyOpenSSL==25.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
service-identity==24.2.0
setuptools==78.1.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
speechmatics-python==3.0.3
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
vanna==0.7.6
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
zope.interface==7.2
